integrate config.json with tests

+++ src/test/conftest.py
@@ -22,8 +22,10 @@ log_helper.setup_for_tests()
 # pylint: enable=wrong-import-position, wrong-import-order
 
 import atexit
+import json
 import logging
 import shutil
+import typing
 
 from nikobot.util import storage
 
@@ -32,36 +34,32 @@ logger = logging.getLogger("test")
 # disable atexit storage saving
 atexit.unregister(storage.PersistentStorage._save_to_disk)
 
-MODULES = ["general", "help", "clear", "music", "avatar", "owner", "tc4.tc4", "mal.malnotifier"]
 STORAGE_DIR = str(pathlib.Path(os.path.dirname(__file__), "..", "..", "test_run").resolve())
-STORAGE_FILE = os.path.join(STORAGE_DIR, "storage.json")
-CACHE_DIR = os.path.join(STORAGE_DIR, "cache")
-TEMP_DIR = os.path.join(STORAGE_DIR, "temp")
+storage.VolatileStorage["config_file"] = "./config.json"
 
-if "test_channel_id" not in storage.VolatileStorage:
-    channel_id_file = str(pathlib.Path(os.path.dirname(__file__), "..", "..", "test_channel_id.txt").resolve())
-    if not os.path.isfile(channel_id_file):
-        raise FileNotFoundError()
-    with open(channel_id_file, "r", encoding="utf8") as file:
-        storage.VolatileStorage["test_channel_id"] = file.readline()
-
-storage.VolatileStorage["test_channel_id"] = 1314744187224195183
-
-storage.VolatileStorage["modules_to_load"] = MODULES
+# load config
+if not os.path.isfile(storage.VolatileStorage["config_file"]):
+    raise FileNotFoundError("Config file couldn't be found")
+with open(storage.VolatileStorage["config_file"], "r", encoding="utf8") as cf:
+    config: dict[str, typing.Any] = json.load(cf)
+if "test" not in config:
+    raise ValueError(f"Missing test section in {storage.VolatileStorage['config_file']}. " +
+                     "You can find an example in config.template.json, located at the repository root.")
 
 # setup testing dirs
 shutil.rmtree(STORAGE_DIR, ignore_errors=True)
 os.makedirs(STORAGE_DIR, exist_ok=True)
 
-storage.VolatileStorage["storage_file"] = STORAGE_FILE
-with open(STORAGE_FILE, "w", encoding="utf8") as f:
+storage.VolatileStorage["storage_file"] = os.path.join(STORAGE_DIR, "storage.json")
+with open(storage.VolatileStorage["storage_file"], "w", encoding="utf8") as f:
     f.write("{}")
 
-storage.VolatileStorage["cache_dir"] = CACHE_DIR
-os.makedirs(CACHE_DIR, exist_ok=True)
+storage.VolatileStorage["cache_dir"] = os.path.join(STORAGE_DIR, "cache")
+os.makedirs(storage.VolatileStorage["cache_dir"], exist_ok=True)
 
-storage.VolatileStorage["temp_dir"] = TEMP_DIR
-os.makedirs(TEMP_DIR, exist_ok=True)
+storage.VolatileStorage["temp_dir"] = os.path.join(STORAGE_DIR, "temp")
+shutil.rmtree(storage.VolatileStorage["temp_dir"], ignore_errors=True)
+os.makedirs(storage.VolatileStorage["temp_dir"], exist_ok=True)
 
 # pylint: disable-next=unused-import
 from .fixtures import setup_storages, bot, testing_bot
